Sphere/transformer_code/src/requirements.txt
transformers>=4.38.0
torch>=2.0.0
pyyaml
